requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
avinit==1.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
confusable-homoglyphs==3.2.0
crispy-bootstrap5==0.6
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.2
django-allauth==0.49.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-model-utils==4.2.0
django-redis==5.2.0
django-registration==3.3
django-sitetree==1.17.3
hiredis==2.0.0
idna==3.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-slugify==6.1.1
python3-openid==3.2.0
pytz==2021.3
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2021.5
urllib3==1.26.9
whitenoise==6.0.0
wrapt==1.14.0
